Move test files around
