+++ src/test/viewer.testy.js
@@ -1,8 +1,8 @@
-import '../test/viewer-setup';
+import './viewer-setup';
 import * as React from 'react';
 import enzyme from 'enzyme';
 import sinon from 'sinon';
-import Viewer from './viewer.jsx';
+import Viewer from '../viewer.jsx';
 //let Viewer = window.gcexports.viewer.Viewer;
 let data = 123;
